refactor(orc8r): S8ProxyResponder servicer as External
