+++ feg/cloud/go/services/feg_relay/feg_relay/main.go
@@ -23,7 +23,7 @@ import (
 	"magma/feg/cloud/go/services/feg_relay"
 	"magma/feg/cloud/go/services/feg_relay/gw_to_feg_relay"
 	nh_servicers "magma/feg/cloud/go/services/feg_relay/gw_to_feg_relay/servicers"
-	"magma/feg/cloud/go/services/feg_relay/servicers"
+	s8proxyResponder_servicers "magma/feg/cloud/go/services/feg_relay/servicers/southbound"
 	lteprotos "magma/lte/cloud/go/protos"
 	"magma/orc8r/cloud/go/service"
 )
@@ -36,7 +36,7 @@ func main() {
 	if err != nil {
 		glog.Fatalf("Error creating Feg Proxy service: %s", err)
 	}
-	servicer, err := servicers.NewFegToGwRelayServer()
+	servicer, err := s8proxyResponder_servicers.NewFegToGwRelayServer()
 
 	if err != nil {
 		glog.Fatalf("Failed to create FegToGwRelayServer: %v", err)
