+++ feg/cloud/go/services/feg_relay/servicers/southbound/terminate_registration.go
